pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-weaviate-service"
version = "2.0.0"
description = "Multi-tenant hybrid-search RAG microservice backed by Weaviate"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Sandesh Neupane"}
]
keywords = ["rag", "weaviate", "hybrid-search", "multi-tenant", "fastapi", "vector-database"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "License :: OSI Approved :: MIT License",
    "Topic :: Database",
]

dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.0",
    "python-dotenv==1.0.1",
    "weaviate-client>=4.9.0,<5.0.0",
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
    "google-genai==0.3.0",
    "requests==2.32.3",
    "loguru==0.7.2",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[project.scripts]
rag-service = "service.main:run_service"

[project.urls]
Homepage = "https://github.com/sandeshneupane/rag-weaviate-service"
Repository = "https://github.com/sandeshneupane/rag-weaviate-service"
Documentation = "https://github.com/sandeshneupane/rag-weaviate-service#readme"

[tool.setuptools]
packages = ["client", "service", "service.core"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]
